+++ liquid/__init__.py
@@ -1,4 +1,11 @@
-__version__ = "0.7.3"
+__version__ = "0.7.4"
+
+try:
+    from markupsafe import escape
+    from markupsafe import Markup
+except ImportError:
+    from liquid.exceptions import escape
+    from liquid.exceptions import Markup
 
 from .mode import Mode
 from .filter import Filter