+++ liquid/builtin/statement.py
@@ -1,16 +1,23 @@
 """Parse tree node and pseudo "tag" for output statements."""
 from typing import TextIO
 
+try:
+    from markupsafe import escape
+    from markupsafe import Markup
+except ImportError:
+    from liquid.exceptions import escape
+    from liquid.exceptions import Markup
 
 from liquid.ast import Node
-from liquid.expression import Expression
 from liquid.context import Context
+from liquid.expression import Expression
 from liquid.lex import tokenize_filtered_expression
-from liquid.stream import TokenStream
-from liquid.tag import Tag
 
 from liquid.parse import parse_filtered_expression
 from liquid.parse import expect
+
+from liquid.stream import TokenStream
+from liquid.tag import Tag
 
 from liquid.token import Token
 from liquid.token import TOKEN_STATEMENT
@@ -33,6 +40,7 @@ class StatementNode(Node):
 
     def render_to_output(self, context: Context, buffer: TextIO):
         val = self.expression.evaluate(context)
+
         if isinstance(val, str):
             # shortcut for common case.
             pass
@@ -41,9 +49,15 @@ class StatementNode(Node):
         elif val is None:
             val = ""
         elif isinstance(val, (list, tuple)):
-            val = "".join(str(v) for v in val)
+            if context.autoescape:
+                val = Markup("").join(_str_if_not(itm) for itm in val)
+            else:
+                val = "".join(_str_if_not(itm) for itm in val)
         else:
             val = str(val)
+
+        if context.autoescape:
+            val = escape(val)
 
         buffer.write(val)
 
@@ -53,10 +67,16 @@ class Statement(Tag):
 
     name = TOKEN_STATEMENT
 
-    def parse(self, stream: TokenStream) -> Node:
+    def parse(self, stream: TokenStream) -> StatementNode:
         tok = stream.current
         expect(stream, TOKEN_STATEMENT)
 
         expr_iter = tokenize_filtered_expression(tok.value)
         node = StatementNode(tok, parse_filtered_expression(TokenStream(expr_iter)))
         return node
+
+
+def _str_if_not(val: object) -> str:
+    if not isinstance(val, str):
+        return str(val)
+    return val